A few changes
This includes all changes that existed before on `from/roles-refactor`
plus a few other changes I introduced myself as suggestions to discuss
when possible:

* Rename `Dictator.Plug.Authorize` to just `Dictator.Plug`. The original
  name felt redundant and verbose when using in multiple controller.
  Since we already know dictator deals with authorization, and there's
  no other plug here (at least for now), this feels correct

* Adds a configurable unauthorized_handler option, which allows
  customizing the http response when this plug halts a response. This is
  necessary if you need custom headers or responses. Two default
  handlers we added: one to keep backward compatibility, and one with
  sane defaults for JSON APIs
