+++ lib/dictator/policies/crud.ex
@@ -1,12 +1,13 @@
-defmodule Dictator.Policies.Standard do
+defmodule Dictator.Policies.CRUD do
   defmacro __using__(opts) do
     quote do
-      use Dictator.Policy, unquote(opts)
+      use Dictator.Policies.Resourceful, unquote(opts)
 
       @foreign_key Keyword.get(unquote(opts), :foreign_key, :user_id)
       @owner_key Keyword.get(unquote(opts), :owner_key, :id)
 
       alias Dictator.Policy
+      alias Dictator.Policies.Resourceful
 
       @impl Policy
       def can?(_, action, _) when action in [:index, :new, :create], do: true
