+++ mix.exs
@@ -34,6 +34,7 @@ defmodule Dictator.MixProject do
 
   defp deps(env) when env in [:dev, :test] do
     [
+      {:dialyxir, "~> 1.0", only: [:dev, :test], runtime: false},
       {:ex_doc, ">= 0.0.0", only: :dev, runtime: false},
       {:credo, "~> 1.2", only: [:dev, :test], runtime: false}
     ]
